필드 정리: DB에 없는 productNo 삭제 및 productOptionNo 중심 설계
타입 통일: 모든 PK/FK 식별자를 Long 타입으로 변경

구조 개선: resultMap과 selectKey를 활용한 표준 매퍼 작성

+++ src/main/java/com/shop/dto/user/review/ReviewSaveRequestDTO.java
@@ -9,9 +9,9 @@ import org.springframework.web.multipart.MultipartFile;
 @Setter
 @ToString
 public class ReviewSaveRequestDTO {
-    private int memberNo;
-    private int productNo;
-    private int orderItemNo;
+    private Long memberNo;       
+    private Long productNo;      
+    private Long orderItemNo;   
     private String title;
     private String content;
     private int rating;
@@ -20,6 +20,7 @@ public class ReviewSaveRequestDTO {
     private Integer userWeight;
     private String sizeRating;
     
+    // 파일 업로드 처리용
     private MultipartFile uploadFile; 
-    private String imageUrl;         
+    private String imageUrl;
 }
